# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CommunityHub",
    "description": "Even better than Meetup",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Community Groups",
      "description": "Groups that meet"
    }
  ],
  "paths": {
    "/community_groups": {
      "get": {
        "tags": ["Community Groups"],
        "summary": "Return all community groups",
        "description": "",
        "operationId": "listCommunityGroups",
        "responses": {
          "200": {
            "description": "Available groups retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityGroupResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Community Groups"],
        "summary": "Add a community group",
        "description": "",
        "operationId": "addCommunityGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityGroupRequestPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityGroupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/community_groups/{id}": {
      "get": {
        "tags": ["Community Groups"],
        "summary": "Get a community group by its id if it exists",
        "operationId": "getCommunityGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityGroupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Community Group not found"
          }
        }
      },
      "put": {
        "tags": ["Community Groups"],
        "summary": "Update a community group by its id",
        "description": "",
        "operationId": "updateCommunityGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityGroupRequestPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityGroupResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Community Groups"],
        "summary": "Delete a community group by its id",
        "operationId": "deleteCommunityGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "id": {
        "name": "id",
        "description": "The Commyunity Group ID",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
    "schemas": {
      "CommunityGroupRequestPost": {
        "type": "object",
        "required": ["name", "location"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Fun Fun Group"
          },
          "description": {
            "type": "string",
            "example": "We do fun fun stuff"
          },
          "location": {
            "type": "string",
            "example": "Toronto, Ontario"
          }
        }
      },
      "CommunityGroupRequestPut": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fun Fun Group"
          },
          "description": {
            "type": "string",
            "example": "We do fun fun stuff"
          },
          "location": {
            "type": "string",
            "example": "Toronto, Ontario"
          }
        }
      },
      "CommunityGroupResponse": {
        "type": "object",
        "required": ["name", "location"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "Fun Fun Group"
          },
          "description": {
            "type": "string",
            "example": "We do fun fun stuff"
          },
          "location": {
            "type": "string",
            "example": "Toronto, Ontario"
          }
        }
      }
    }
  }
}